api/main_simple.py
#!/usr/bin/env python3
"""
Simplified Universal Knowledge Hub API for Python 3.13.5
"""

import asyncio
import logging
import time
import uuid
from contextlib import asynccontextmanager
from typing import Any, Optional
from datetime import datetime, timedelta

import uvicorn
from fastapi import FastAPI, HTTPException, Request, Response, Depends
from fastapi.middleware.cors import CORSMiddleware
from fastapi.responses import JSONResponse
from pydantic import ValidationError, Field, BaseModel

from dotenv import load_dotenv
import os

# Load environment variables
load_dotenv()

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Pydantic models
class QueryRequest(BaseModel):
    query: str = Field(..., min_length=1, max_length=10000, description="The user's question or query")
    max_tokens: Optional[int] = Field(1000, ge=100, le=4000, description="Maximum tokens for response")
    confidence_threshold: Optional[float] = Field(0.8, ge=0.0, le=1.0, description="Minimum confidence score")

class QueryResponse(BaseModel):
    answer: str = Field(..., description="The AI-generated answer")
    confidence: float = Field(..., ge=0.0, le=1.0, description="Confidence score")
    query_id: str = Field(..., description="Unique query identifier")
    processing_time: Optional[float] = Field(None, description="Processing time in seconds")
    metadata: dict[str, Any] = Field(default_factory=dict, description="Additional metadata")

class HealthResponse(BaseModel):
    status: str = Field(..., description="Service status")
    version: str = Field(..., description="API version")
    timestamp: float = Field(..., description="Current timestamp")
    uptime: float = Field(..., description="Service uptime in seconds")

class AuthRequest(BaseModel):
    username: str = Field(..., description="Username")
    password: str = Field(..., description="Password")

class AuthResponse(BaseModel):
    access_token: str = Field(..., description="JWT access token")
    token_type: str = Field(..., description="Token type")
    user_id: str = Field(..., description="User ID")
    role: str = Field(..., description="User role")

# Global variables
startup_time = time.time()
app_version = "1.0.0"

@asynccontextmanager
async def lifespan(app: FastAPI):
    """Application lifespan manager."""
    global startup_time
    
    # Startup
    startup_time = time.time()
    logger.info("🚀 Starting Universal Knowledge Hub - Python 3.13.5 Compatible")
    
    yield
    
    # Shutdown
    logger.info("🛑 Shutting down Universal Knowledge Hub")

# Create FastAPI app
app = FastAPI(
    title="Universal Knowledge Hub",
    description="AI-powered knowledge platform - Python 3.13.5 Compatible",
    version=app_version,
    lifespan=lifespan
)

# Add CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Request ID middleware
@app.middleware("http")
async def add_request_id(request: Request, call_next):
    request_id = str(uuid.uuid4())
    request.state.request_id = request_id
    response = await call_next(request)
    response.headers["X-Request-ID"] = request_id
    return response

# Global exception handler
@app.exception_handler(Exception)
async def global_exception_handler(request: Request, exc: Exception):
    logger.error(f"Unhandled exception: {exc}")
    return JSONResponse(
        status_code=500,
        content={
            "error": "Internal server error",
            "request_id": getattr(request.state, "request_id", "unknown"),
            "timestamp": time.time()
        }
    )

@app.exception_handler(ValidationError)
async def validation_exception_handler(request: Request, exc: ValidationError):
    return JSONResponse(
        status_code=422,
        content={
            "error": "Validation error",
            "details": exc.errors(),
            "request_id": getattr(request.state, "request_id", "unknown")
        }
    )

# Root endpoint
@app.get("/", response_model=dict[str, Any])
async def root():
    """Root endpoint with API information."""
    return {
        "message": "Universal Knowledge Hub API",
        "version": app_version,
        "python_version": "3.13.5",
        "status": "running",
        "timestamp": time.time()
    }

# Health check endpoint
@app.get("/health", response_model=HealthResponse)
async def health_check():
    """Health check endpoint."""
    uptime = time.time() - startup_time
    return HealthResponse(
        status="healthy",
        version=app_version,
        timestamp=time.time(),
        uptime=uptime
    )

# Query endpoint
@app.post("/query", response_model=QueryResponse)
async def process_query(request: QueryRequest):
    """Process a knowledge query."""
    start_time = time.time()
    query_id = str(uuid.uuid4())
    
    # Simulate AI processing
    await asyncio.sleep(0.1)  # Simulate processing time
    
    # Generate mock response
    answer = f"Here is the answer to your query: '{request.query}'. This is a simulated response from the Universal Knowledge Hub running on Python 3.13.5."
    confidence = 0.95
    processing_time = time.time() - start_time
    
    return QueryResponse(
        answer=answer,
        confidence=confidence,
        query_id=query_id,
        processing_time=processing_time,
        metadata={
            "model": "simulated-ai",
            "python_version": "3.13.5",
            "request_tokens": len(request.query.split())
        }
    )

# Authentication endpoints
@app.post("/auth/login", response_model=AuthResponse)
async def login(auth_request: AuthRequest):
    """Login endpoint."""
    # Simulate authentication
    if auth_request.username == "admin" and auth_request.password == "password":
        return AuthResponse(
            access_token="mock-jwt-token",
            token_type="bearer",
            user_id="admin-123",
            role="admin"
        )
    else:
        raise HTTPException(status_code=401, detail="Invalid credentials")

# Metrics endpoint
@app.get("/metrics")
async def get_metrics():
    """Get system metrics."""
    return {
        "uptime": time.time() - startup_time,
        "version": app_version,
        "python_version": "3.13.5",
        "status": "healthy"
    }

# Test endpoint
@app.get("/test")
async def test_endpoint():
    """Test endpoint for Python 3.13.5 compatibility."""
    return {
        "message": "Python 3.13.5 compatibility test successful!",
        "features": [
            "Modern typing syntax",
            "Async/await support",
            "FastAPI integration",
            "Pydantic validation"
        ],
        "timestamp": time.time()
    }

if __name__ == "__main__":
    print("🚀 Starting Universal Knowledge Hub - Python 3.13.5 Compatible")
    print("✅ All compatibility fixes applied")
    print("🌐 Server will be available at: http://127.0.0.1:8000")
    print("📋 Available endpoints:")
    print("   - GET / - Root endpoint")
    print("   - GET /health - Health check")
    print("   - POST /query - Process queries")
    print("   - POST /auth/login - Authentication")
    print("   - GET /metrics - System metrics")
    print("   - GET /test - Compatibility test")
    
    uvicorn.run(app, host="127.0.0.1", port=8000) 