requirements-dev.txt
# Development dependencies
# Code formatting and linting
black==25.1.1
flake8==7.3.0
mypy==1.13.0
mypy-extensions==1.1.0
isort==5.13.2
autopep8==2.0.4

# Development tools
ipython==8.18.1
ipdb==0.13.13
pre-commit==3.6.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Type stubs
types-requests==2.31.0.10
types-redis==4.6.0.11
types-python-dateutil==2.8.19.14 
# Universal Knowledge Platform - Backend Requirements
# Core dependencies for the multi-agent RAG system
# For full entity extraction, run: python -m spacy download en_core_web_sm

# Core FastAPI and web framework
fastapi==0.116.1
uvicorn==0.35.0
starlette==0.47.2

# Data validation and settings
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2

# Environment and configuration
python-dotenv==1.1.1

# Security and authentication
bcrypt==4.3.0
cryptography==45.0.5
PyJWT==2.10.1
passlib==1.7.4
email-validator==2.2.0

# Logging and monitoring
structlog==25.4.0
prometheus_client==0.22.1

# HTTP and async utilities
aiohttp==3.12.14
requests==2.32.4
tenacity==9.1.2
python-multipart==0.0.20

# Data processing
numpy==2.0.2
spacy==3.7.4

# Caching and storage
aiocache==0.12.3
aioredis==2.0.1
redis==5.0.1

# Content sanitization
bleach==6.2.0

# System monitoring
psutil==7.0.0

# OpenTelemetry for distributed tracing
opentelemetry-api==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-instrumentation==0.56b0
opentelemetry-instrumentation-fastapi==0.56b0
opentelemetry-instrumentation-logging==0.56b0
opentelemetry-instrumentation-asgi==0.56b0
opentelemetry-semantic-conventions==0.56b0
opentelemetry-util-http==0.56b0

# Database and search (optional - for future use)
SQLAlchemy==2.0.41
elasticsearch==8.15.0

# Development and testing
pytest==8.4.1
pytest-asyncio==1.1.0
redis
async
asyncio
aioredis